new DB tests, changes CHID of user to long long int, db access

+++ models/user.cpp
@@ -4,7 +4,7 @@
 
 #include "user.h"
 
-user::user(std::string name, unsigned long int uid, unsigned long int chid, bool active) {
+user::user(std::string name, unsigned long int uid, long long int chid, bool active) {
     this->active = active;
     this->chid = chid;
     this->uid = uid;
